move cicle class to different file and rename circleutils.h to circlefit.h
