+++ DFTFringe_Dale.pro
@@ -240,8 +240,7 @@ HEADERS  += mainwindow.h \
     myutils.h \
     pixelstats.h \
     utils.h \
-    circleutils.h \
-    circle.h \
+    circlefit.h \
     astigstatsdlg.h \
     averagewavefrontfilesdlg.h \
     astigscatterplot.h \
