+++ circlefit.h
@@ -1,11 +1,39 @@
-#ifndef CIRCLEUTILS_H
-#define CIRCLEUTILS_H
+#ifndef CIRCLEFIT_H
+#define CIRCLEFIT_H
 #include <iostream>
 #include <cmath>
 #include <limits>
 #include <iomanip>
 #include <cstdlib>
-#include "circle.h"
+
+/************************************************************************
+            DECLARATION OF THE CLASS CIRCLE
+************************************************************************/
+// Class for Circle
+// A circle has 7 fields:
+//     a, b, r (of type reals), the circle parameters
+//     s (of type reals), the estimate of sigma (standard deviation)
+//     g (of type reals), the norm of the gradient of the objective function
+//     i and j (of type int), the iteration counters (outer and inner, respectively)
+
+class Circle
+{
+public:
+
+    // The fields of a Circle
+    double a, b, r, s, g, Gx, Gy;
+    int i, j;
+
+    // constructors
+    Circle();
+    Circle(double aa, double bb, double rr);
+
+    // routines
+    void print(void);
+
+    // no destructor we didn't allocate memory by hand.
+};
+
 
 class CircleData
 {
@@ -47,4 +75,5 @@ Circle CircleFitByKasa (CircleData& data);
 Circle CircleFitByPratt (CircleData& data);
 Circle CircleFitByTaubin (CircleData& data);
 int CircleFitByLevenbergMarquardtFull (const CircleData& data, const Circle& circleIni, double LambdaIni, Circle& circle);
-#endif // CIRCLEUTILS_H
+
+#endif // CIRCLEFIT_H
