+++ surfacemanager.cpp
@@ -63,8 +63,7 @@
 #include "foucaultview.h"
 #include "myutils.h"
 #include "contourview.h"
-#include "circleutils.h"
-#include "circle.h"
+#include "circlefit.h"
 #include "utils.h"
 #include "wavefrontfilterdlg.h"
 #include "reportdlg.h"
